requirements.txt
python-dotenv==1.0.1
spotipy==2.22.1
gspread==6.0.0